Develop

+++ src/app/api/comment/delete/route.ts
@@ -1,25 +1,33 @@
 import prisma from "@/lib/prisma";
 import { NextRequest, NextResponse } from "next/server";
-import jwt from "jsonwebtoken";
+import { DeleteSchema } from "@/utils/validation";
+import z from "zod";
+import { loginCheck } from "@/utils/loginCheck";
 
 export async function POST(req: NextRequest) {
 
-    const { id } = await req.json();
-    const JWT_SECRET = process.env.JWT_SECRET;
+    const { userId, id } = await req.json();
 
     try {
         // ログインしているかどうかの判定
+        const isLogin = await loginCheck(req);
 
-        const token = req.cookies.get("auth_token")?.value;
-        const data = await jwt.verify(token!, JWT_SECRET!);
+        if (!isLogin) {
+            return NextResponse.json({message: "ログインしていません。", success: false, login: false}, {status: 401});
+        }
+        
+        // 削除時バリデーションチェック
+        const issue = DeleteSchema.safeParse({ userId, id });
 
-        if (!data) {
-            return NextResponse.json({ message: "ログインしていません。", success: false }, { status: 500 });
+        if (!issue.success) {
+            const validation = z.flattenError(issue.error);
+            const message = validation.fieldErrors ?? null;
+            return NextResponse.json({ message: message, success: false }, { status: 400 });
         }
 
-        if (id) {
+        if (id && userId) {
             await prisma.comments.delete({
-                where: { id: id }
+                where: { id: Number(id), userId: Number(userId) }
             });
             return NextResponse.json({ message: "コメントを削除しました。", success: true }, { status: 200 });
         } else {
